requirements.txt
Django==3.2.6
psycopg2-binary==2.9.1
whitenoise==5.3.0
requests==2.24.0
djangorestframework==3.12.4
Markdown==3.1.1
django-filter==2.1.0
django-inlinecss==0.3.0
python-dateutil==2.8.1
django-cors-headers==3.2.0
wheel==0.32.3
gunicorn==19.9.0
python-dotenv~=0.15.0
dataclasses==0.6
